requirements.txt
# Core
numpy
pandas
matplotlib
seaborn
python-dotenv

# ML (Random Forest, metrics, preprocessing)
scikit-learn

# Data
alpaca-trade-api
alpaca-py

# Deep learning (LSTM)
tensorflow-macos; platform_system=="Darwin"
tensorflow; platform_system!="Darwin"

# etc
joblib
scipy